% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFuncs.R
\name{calculate_qvalues}
\alias{calculate_qvalues}
\title{Calculate FDRs for a given vector of p-values with qvalue, fallback to BH
adjustment if qvalue fails. Default to BH in case of less than 200 p-values.}
\usage{
calculate_qvalues(pvals, stat_test = "limma")
}
\arguments{
\item{pvals}{A numeric vector of p-values for which to calculate FDRs.}

\item{stat_test}{A character string indicating the statistical test used to
generate the p-values.}
}
\value{
A numeric vector of q-values corresponding to the input p-values.
}
\description{
Calculate FDRs for a given vector of p-values with qvalue, fallback to BH
adjustment if qvalue fails. Default to BH in case of less than 200 p-values.
}
\examples{
pvals <- c(0.01, 0.04, 0.05, 0.20, 0.50, NA)
qvals <- calculate_qvalues(pvals)
print(qvals)
}
