% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFuncs.R
\name{check_for_polystest}
\alias{check_for_polystest}
\title{Check SummarizedExperiment for PolySTest Requirements}
\usage{
check_for_polystest(se)
}
\arguments{
\item{se}{A SummarizedExperiment object.}
}
\value{
Invisible TRUE if checks pass; otherwise, warnings or errors are
thrown.
}
\description{
Performs checks on a SummarizedExperiment object to ensure it is properly
formatted for PolySTest analysis. It checks for specific metadata properties,
the number of assays, and the distribution of conditions and replicates.
}
\examples{
data(liver_example)
check_for_polystest(liver_example)

}
