% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFuncs.R
\name{FindFCandQlim}
\alias{FindFCandQlim}
\title{Find Optimal Fold-Change and Q-Value Thresholds}
\usage{
FindFCandQlim(Qvalue, LogRatios)
}
\arguments{
\item{Qvalue}{A matrix of q-values for the test results from Unpaired or Paired}

\item{LogRatios}{A matrix of log ratios.}
}
\value{
A numeric vector containing the optimal fold-change threshold and q-value threshold.
}
\description{
This function calculates the "optimal" fold-change and q-value thresholds by maximizing the
percental output of features commonly found for limma, rank products, permutation tests, and NA tests.
It sets any NA values in the Qvalue matrix to 1, determines the smallest q-value, selects a range
of q-values around it, runs over different fold-change (FC) thresholds, modifies the Qvalue matrix
based on the FC threshold, calculates the distribution of significant features for each q-value
threshold, and returns the combination of FC and q-value thresholds that result in the highest mean
distribution of significant features.
TODO: adapt to output of Paired/Unpaired functions
}
\examples{
Qvalue <- matrix(c(0.01, 0.02, 0.03, 0.04, 0.05), nrow = 5, ncol = 3)
LogRatios <- matrix(c(1.2, 0.8, 1.5, -0.5, 0.2, 0.9, -1.1, 0.7, 1.8, -0.9, 0.3, 1.1), nrow = 4, ncol = 3)
thresholds <- FindFCandQlim(Qvalue, LogRatios)
print(thresholds)

}
\keyword{internal}
