% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFuncs.R
\name{get_numthreads}
\alias{get_numthreads}
\title{Set number of threads (default is 4)}
\usage{
get_numthreads()
}
\description{
This function sets the number of threads for parallel processing. If the
environment variable SHINY_THREADS is set, the number of threads is set
to the value of SHINY_THREADS.
}
