% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatTestUnpaired.R
\name{limma_unpaired}
\alias{limma_unpaired}
\title{Perform unpaired limma analysis
This function performs unpaired limma analysis on Data.}
\usage{
limma_unpaired(Data, NumCond, NumReps, RRCateg)
}
\arguments{
\item{Data}{A matrix of gene expression data.}

\item{NumCond}{The number of conditions in the experiment.}

\item{NumReps}{The number of replicates per condition.}

\item{RRCateg}{A matrix specifying the coniditons to be compared}
}
\value{
A list containing the following results:
\itemize{
\item plvalues: The p-values from limma tests.
\item qlvalues: The q-values from limma tests.
\item Sds: The standard deviations of the Bayesian linear model.
@details This function performs unpaired limma analysis on Data. It calculates the p-values and q-values for each row, indicating the significance of the difference between the two datasets.
}
}
\description{
Perform unpaired limma analysis
This function performs unpaired limma analysis on Data.
}
\examples{
# Example usage of limma_unpaired function
data <- matrix(rnorm(1500), ncol = 15)
RRCateg <- matrix(c(1, 2, 2, 3), nrow = 2, ncol = 2)
result <- limma_unpaired(data, 3, 5, RRCateg)
print(head(result))
}
\keyword{analysis}
\keyword{limma}
\keyword{unpaired}
