% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liver-example.R
\docType{data}
\name{liver_example}
\alias{liver_example}
\title{Example data set liver_example for PolySTest}
\format{
A SummarizedExperiment object
}
\source{
Protein expressions from the livers of mice fed with different diets: high fat,
TTA component, Fish oil and TTA + fish oil
}
\usage{
liver_example
}
\description{
This data set is a SummarizedExperiment objects containing liver proteomics
of mice fed with four different diets. The data contains 3 replicates per
diet (condition). The rowData of the object contains the output from running
the PolySTest statistical tests on the data.
}
\keyword{datasets}
