% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFuncs.R
\name{MissValPDistr}
\alias{MissValPDistr}
\title{Calculate the distribution of missing values}
\usage{
MissValPDistr(NumReps, PercNA)
}
\arguments{
\item{NumReps}{An integer indicating the number of repetitions.}

\item{PercNA}{A numeric value indicating the percentage of missing values.}
}
\value{
A numeric vector containing the distribution of missing values.
}
\description{
This function calculates the distribution of missing values for a given
number of repetitions and percentage of missing values.
}
\examples{
MissValPDistr(10, 0.2)
}
