% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatTestPaired.R
\name{Paired}
\alias{Paired}
\title{Paired Test Analysis}
\usage{
Paired(MAData, NumCond, NumReps)
}
\arguments{
\item{MAData}{A matrix containing the data for the analysis.}

\item{NumCond}{The number of conditions in the analysis.}

\item{NumReps}{The number of replicates per condition.}
}
\value{
A list containing the results of the paired test analysis, including p-values, q-values, and ratios.
}
\description{
This function performs paired test analysis on the given MAData.
}
\examples{
MAData <- matrix(rnorm(100), ncol = 5)
results <- Paired(MAData, 5, 2)
print(results)

}
