% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatTestPaired.R
\name{PolySTest_paired}
\alias{PolySTest_paired}
\title{PolySTest for paired tests}
\usage{
PolySTest_paired(fulldata, allComps)
}
\arguments{
\item{fulldata}{A SummarizedExperiment or derived object that contains
the quantitative data as required for PolySTest}

\item{allComps}{A matrix containing the reference matrix specifying the
pairs of conditions to compare (each comparison given as separate row)}
}
\value{
SummarizedExperiment with added columns for p-values and q-values
in rowData
}
\description{
Combining the power of different statistical tests
}
\examples{
# Assuming 'fulldata' is a SummarizedExperiment object
# and 'allComps' is a matrix containing the reference matrix
# specifying the pairs of conditions to compare
PolySTest_paired(fulldata, allComps)

}
