% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFuncs.R
\name{prepare_output_data}
\alias{prepare_output_data}
\title{Prepare Output Data for PolySTest Results}
\usage{
prepare_output_data(fulldata, Pvalue, Qvalue, LogRatios, testNames, allComps)
}
\arguments{
\item{fulldata}{A SummarizedExperiment object containing the initial dataset.}

\item{Pvalue}{A matrix of p-values from the statistical tests.}

\item{Qvalue}{A matrix of q-values corresponding to the p-values.}

\item{LogRatios}{A matrix of log-ratio values for the comparisons.}

\item{testNames}{A vector of names for each of the statistical tests performed.}

\item{allComps}{A matrix specifying the pairs of conditions compared,
each row represents a pair.}
}
\value{
The updated SummarizedExperiment object with additional columns in
rowData for log-ratios, p-values, and q-values.
}
\description{
This function processes the results of PolySTest, including log-ratios, p-values,
and q-values for each statistical test applied, and integrates them into the
rowData of the provided SummarizedExperiment object. It optionally handles
separate t-test q-values and unifies q-values across multiple tests.
}
\details{
This function first checks for the presence of t-test results within
the provided data, segregates them if present, and then optionally
unifies q-values across tests if more than one test is specified.
It then organizes and renames the matrices of log-ratios, p-values,
and q-values according to the comparisons and tests performed, and
merges these matrices into the rowData of the provided
SummarizedExperiment object. Finally, it prints a summary of the
number of features with FDR < 0.01 for each test and comparison.
}
\examples{
# Assuming 'fulldat' is your SummarizedExperiment object, 'Pvalue', 'Qvalue',
# and 'LogRatios' are matrices of your test results, 'testNames' is your
# vector of test names, and 'allComps' specifies your condition pairs:
# fulldat <- prepare_output_data(fulldat, Pvalue, Qvalue, LogRatios, testNames, allComps)

}
