% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFuncs.R
\name{StatsForPermutTest}
\alias{StatsForPermutTest}
\title{Calculate statistics for permutation test}
\usage{
StatsForPermutTest(Data, Paired)
}
\arguments{
\item{Data}{A matrix or data frame containing the data for the test.}

\item{Paired}{A logical value indicating whether the test is paired or not.}
}
\value{
A numeric vector containing the calculated statistics.
}
\description{
This function calculates the statistics for a permutation test based on the
input data.
}
\examples{
Data <- matrix(rnorm(100), ncol = 10)
StatsForPermutTest(Data, Paired = FALSE)
}
