% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatTestPaired.R
\name{ttest_paired}
\alias{ttest_paired}
\title{Perform paired t-tests}
\usage{
ttest_paired(tMAData)
}
\arguments{
\item{tMAData}{A matrix of data for running row-wise t-tests}
}
\value{
A list containing the p-values and q-values (qvalue package)
}
\description{
This function performs row-wise paired t-tests on a data frame containing
log-fold changes
}
\examples{
tMAData <- matrix(rnorm(1000), nrow = 100)
tout <- ttest_paired(tMAData)
head(tout$qtvalues)
}
\keyword{analysis}
\keyword{paired}
\keyword{t-test}
