% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatTestUnpaired.R
\name{ttest_unpaired}
\alias{ttest_unpaired}
\title{Perform unpaired t-tests on two datasets}
\usage{
ttest_unpaired(tData, trefData)
}
\arguments{
\item{tData}{The first dataset, a matrix or data frame}

\item{trefData}{The second dataset, a matrix or data frame}
}
\value{
A list containing the p-values and q-values for each row
}
\description{
This function performs unpaired t-tests between corresponding rows of two datasets.
It calculates the p-values and q-values for each row, indicating the significance of the difference between the two datasets.
}
\examples{
tData <- matrix(rnorm(100), nrow = 10)
trefData <- matrix(rnorm(100), nrow = 10)
result <- ttest_unpaired(tData, trefData)
print(result$ptvalues)
print(result$qtvalues)

}
