% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatTestUnpaired.R
\name{Unpaired}
\alias{Unpaired}
\title{Unpaired}
\usage{
Unpaired(Data, NumCond, NumReps)
}
\arguments{
\item{Data}{A matrix of gene expression data.}

\item{NumCond}{The number of conditions.}

\item{NumReps}{The number of replicates per condition.}
}
\value{
A list containing various results of the significance analysis, including lratios,
ptvalues, plvalues, pRPvalues, pPermutvalues, qtvalues, qlvalues, qRPvalues, qPermutvalues, and Sds.
}
\description{
This function performs significance analysis for unpaired data.
}
\examples{
Data <- matrix(rnorm(100), ncol = 5)
result <- Unpaired(Data, 5, 2)
print(result)
}
