% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatTestUnpaired.R
\name{UnpairedDesign}
\alias{UnpairedDesign}
\title{UnpairedDesign function}
\usage{
UnpairedDesign(Data, RRCateg, NumCond, NumReps)
}
\arguments{
\item{Data}{A matrix of gene expression data.}

\item{RRCateg}{A matrix of two rows specifying the condition number(s) to be compared (first row versus second row)}

\item{NumCond}{The number of conditions in the experiment.}

\item{NumReps}{The number of replicates per condition.}
}
\value{
A list containing the following results:
\itemize{
\item lratios: The log ratios of gene expression between conditions.
\item ptvalues: The p-values from t-tests.
\item plvalues: The p-values from limma tests.
\item pRPvalues: The p-values from rank products tests.
\item pPermutvalues: The p-values from permutation tests.
\item qtvalues: The q-values from t-tests.
\item qlvalues: The q-values from limma tests.
\item qRPvalues: The q-values from rank products tests.
\item qPermutvalues: The q-values from permutation tests.
\item Sds: The standard deviations of the Bayesian linear model.
}
}
\description{
This function performs significance analysis using different statistical tests such as limma, rank products, and permutation tests.
}
\examples{
# Example usage of UnpairedDesign function
data <- matrix(rnorm(1500), ncol = 15)
RRCateg <- matrix(c(1, 2, 2, 3), nrow = 2, ncol = 2)
result <- UnpairedDesign(data, RR, 3, 5)
print(result)

}
