% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatTestUnpaired.R
\name{perm_unpaired}
\alias{perm_unpaired}
\title{perm_unpaired function}
\usage{
perm_unpaired(tData, trefData)
}
\arguments{
\item{tData}{The test data matrix.}

\item{trefData}{The reference data matrix.}
}
\value{
A list containing the p-values and q-values for the permutation test.
}
\description{
This function performs permutation testing for unpaired data.
}
\details{
The function adds columns from the randomized full set to reach the minimum number of
permutation columns (NumPermCols) replicates. It randomizes the sign as well to avoid
tendencies to one or the other side. In the unpaired case, it also normalizes by the
mean of the entire sample to avoid strange effects. The function then performs permutation
testing using parallel computing, and calculates the p-values and q-values based on the permutation results.
}
\examples{
tData <- matrix(rnorm(1000), nrow = 100)
trefData <- matrix(rnorm(1000), nrow = 100)
result <- perm_unpaired(tData, trefData)

}
