% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations.R
\name{plotUpset}
\alias{plotUpset}
\title{Plot UpSet}
\usage{
plotUpset(fulldata, qlim = 0.05, fclim = c(0, 0))
}
\arguments{
\item{fulldata}{A \code{SummarizedExperiment} object containing the data.}

\item{qlim}{A numeric value, the q-value threshold for significance.}

\item{fclim}{A numeric vector, specifying fold change limits for filtering.}
}
\value{
An UpSet plot visualizing the intersections of significant features.
}
\description{
Visualizes the intersections of significant features across multiple comparisons
using an UpSet plot. Summarizes all comparisons from all tests
}
\examples{
data(liver_example)

plotUpset(liver_example, qlim = 0.05)

}
