% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatTestUnpaired.R
\name{rp_unpaired}
\alias{rp_unpaired}
\title{rp_unpaired function}
\usage{
rp_unpaired(tData, trefData)
}
\arguments{
\item{tData}{The data matrix for the test group.}

\item{trefData}{The data matrix for the reference group.}

\item{NumReps}{The number of replicates.}
}
\value{
A list containing the p-values and q-values.
}
\description{
This function calculates the p-values and q-values for unpaired random pairing combinations.
}
\examples{
tData <- matrix(rnorm(1000), nrow = 100)
trefData <- matrix(rnorm(1000), nrow = 100)
rp_unpaired(tData, trefData)

}
