% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFuncs.R
\name{update_conditions_with_lcs}
\alias{update_conditions_with_lcs}
\title{Update Conditions with Longest Common Subsequence}
\usage{
update_conditions_with_lcs(fulldata, default = NULL)
}
\arguments{
\item{fulldata}{A SummarizedExperiment or derived object that contains
the quantitative data as specified for PolySTest}

\item{default}{A vector of the length of the number of conditions suggesting
their names}
}
\value{
Summariz3dExperiment with updated conditions
}
\description{
This function iterates over a set of conditions and updates each condition
with the longest common subsequence of column names associated with that condition.
}
\examples{
se <- SummarizedExperiment(assays = list(counts = matrix(rnorm(200), ncol = 10)))
metadata(se) <- list(NumCond = 2, NumReps = 5)
colnames(assay(se)) <- c(paste("CondA_Rep", 1:5, sep = ""),
                         paste("CondB_Rep", 1:5, sep = ""))
default_conditions <- c("Condition_A", "Condition_B")
updated_conditions <- update_conditions_with_lcs(se, default_conditions)
print(colData(updated_conditions))
}
