% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFuncs.R
\name{update_conditions_with_lcs}
\alias{update_conditions_with_lcs}
\title{Update Conditions with Longest Common Subsequence}
\usage{
update_conditions_with_lcs(dat, numCond, numReps, conditions)
}
\arguments{
\item{dat}{A data frame or matrix with column names.}

\item{numCond}{The number of conditions to process.}

\item{numReps}{The number of replicates for each condition.}

\item{conditions}{A vector of condition names to be updated.}
}
\value{
Updated vector of condition names.
}
\description{
This function iterates over a set of conditions and updates each condition
with the longest common subsequence of column names associated with that condition.
}
\examples{
dat <- matrix(ncol = 6, nrow = 10)
colnames(dat) <- c("Cond1_Rep1", "Cond1_Rep2", "Cond2_Rep1", "Cond2_Rep2", "Cond3_Rep1", "Cond3_Rep2")
conditions <- c("Cond1", "Cond2", "Cond3")
update_conditions_with_lcs(dat, 3, 2, conditions)
}
